Made bezier_curve.py
Create a smooth line for animation with bezier curve
This also can use easing/inbetweening methods to start and end the animation in a certain way.
Numpy is needed for this, but this is just used for the speed. The script could be changed so the builtin math module can be used.
This script in combination with pynput.mouse could be used to create human like mouse movement. Just add a sleep in loop.